feat: x402 refactor in preparation for middleware
