+++ typescript/packages/x402/src/shared/evm/index.ts
@@ -1,0 +1,1 @@
+export * from "./usdc";
