+++ typescript/packages/x402/src/shared/evm/usdc.ts
@@ -1,7 +1,7 @@
-import { Account, Address, Chain, Client, Hex, PublicClient, Transport } from "viem";
-import { usdcABI as abi } from "./erc20PermitABI";
-import { config } from "./config";
-import { ConnectedClient } from "./wallet";
+import { Account, Address, Chain, Client, Transport } from "viem";
+import { config } from "../../types/shared/evm/config";
+import { usdcABI as abi } from "../../types/shared/evm/erc20PermitABI";
+import { ConnectedClient } from "../../types/shared/evm/wallet";
 
 export function getUsdcAddress<
   transport extends Transport,
