+++ typescript/packages/x402/src/types/shared/evm/index.ts
@@ -1,5 +1,4 @@
 export * from "./config";
 export * from "./eip3009";
 export * from "./erc20PermitABI";
-export * from "./usdc";
 export * from "./wallet";
