+++ typescript/packages/x402/src/types/verify/index.ts
@@ -1,1 +1,2 @@
 export * from "./x402Specs";
+export * from "./facilitator";
