+++ typescript/packages/x402/tsup.config.ts
@@ -4,8 +4,8 @@ const baseConfig = {
   entry: {
     'index': 'src/index.ts',
     'shared/index': 'src/shared/index.ts',
+    'shared/evm/index': 'src/shared/evm/index.ts',
     'client/index': 'src/client/index.ts',
-    'facilitator/index': 'src/facilitator/index.ts',
     'verify/index': 'src/verify/index.ts',
     'types/index': 'src/types/index.ts'
   },
@@ -13,8 +13,6 @@ const baseConfig = {
     resolve: true,
   },
   sourcemap: true,
-  outDir: 'dist/esm',
-  clean: true,
   target: 'node16'
 }
